==> 1_data_clean_shelter.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np
import re

# Import Data

austin_intakes = pd.read_csv('./datasets/raw_data/austin_animal_center_intakes.csv')

austin_intakes.head(2)

austin_outcomes = pd.read_csv('./datasets/raw_data/austin_animal_center_outcomes.csv')

austin_outcomes.head(2)

austin_outcomes_2 = pd.read_csv('./datasets/raw_data/austin_animal_center_outcomes_2.csv')

austin_outcomes_2.head(2)

# Clean Data

## Austin Intakes

# Alternative animals
austin_intakes[austin_intakes['Animal Type']!='Dog']['Animal Type'].value_counts()

austin_intakes.shape

austin_intakes[austin_intakes['Animal Type']=='Dog'].head(3)

Drop non-dogs

austin_intakes = austin_intakes[austin_intakes['Animal Type']=='Dog'].copy()

austin_intakes.shape

## Austin Outcomes

### Merge outcomes .csv

Merge austin outcomes data sets

austin_outcomes = pd.concat([austin_outcomes, austin_outcomes_2])

Drop duplicated rows

austin_outcomes.drop_duplicates(keep='first', inplace=True)

austin_outcomes.shape

Inspect non-dogs

austin_outcomes[austin_outcomes['Animal Type']!='Dog']['Animal Type'].value_counts()

Drop non-dogs

austin_outcomes = austin_outcomes[austin_outcomes['Animal Type']=='Dog'].copy()

## Merge Intakes & Outcomes

austin_intakes.shape

austin_outcomes.shape

View duplicate intake Animal IDs

austin_intakes[austin_intakes['Animal ID'].duplicated(keep=False)==True].sort_values(by='Animal ID').head()

Duplciates appear to be same animal passing through intake at the shelter multiple times

austin_outcomes[austin_outcomes['Animal ID'].duplicated(keep=False)==True].sort_values(by='Animal ID').head()

austin_intakes.shape

austin_outcomes.shape


Rows of austin intakes and austin outcomes appear to largely line up, however dates for each animal are not necessarily listed in the same order and intakes has ~40 more rows

austin_outcomes.sort_values(by='Animal ID').head()

austin_intakes.sort_values(by='Animal ID').head()

Identify missing outcome Animal IDs

in_id = set(austin_intakes['Animal ID'])
out_id = set(austin_outcomes['Animal ID'])

len(list(out_id.difference(in_id)))

Missing intake data on 381 Animal IDs from outcomes

List of all Animal IDs in austin outcomes and intakes

austin_in_out_ids = list(in_id.union(out_id))

A few dogs appear in the system many times :(

austin_intakes['Animal ID'].value_counts().sort_values(ascending=False).head(10)

Convert DateTimes

austin_intakes['DateTime'] = pd.to_datetime(austin_intakes['DateTime'])

austin_intakes.sort_values(by='DateTime').head(3)

austin_outcomes['DateTime'] = pd.to_datetime(austin_outcomes['DateTime'])

austin_outcomes.head(3)

Merge Intake & Outcomes

austin_intakes.head(3)

austin_intakes.drop(columns=['MonthYear', 'Animal Type'], inplace=True)
austin_outcomes.drop(columns=['MonthYear', 'Animal Type'], inplace=True)

Sort intakes & outcomes

austin_intakes.sort_values(by=['Animal ID', 'DateTime'], inplace=True)
austin_intakes.reset_index(drop=True, inplace=True)

austin_outcomes.sort_values(by=['Animal ID', 'DateTime'], inplace=True)
austin_outcomes.reset_index(drop=True, inplace=True)

austin_intakes.head(3)

austin_outcomes.head(3)

Count ins and outs

in_count = austin_intakes['Animal ID'].value_counts()
out_count = austin_outcomes['Animal ID'].value_counts()
repeat_in = in_count[in_count > 1]
repeat_out = out_count[out_count > 1]

Separate repeat intakes from singular instakes

repeat_in = austin_intakes[austin_intakes['Animal ID'].isin(list(repeat_in.index))].copy(deep=True)
solo_in = austin_intakes[~austin_intakes['Animal ID'].isin(repeat_in['Animal ID'].unique())].copy(deep=True)

Separate repeat outcomes from singular outcomes

repeat_out = austin_outcomes[austin_outcomes['Animal ID'].isin(list(repeat_out.index))].copy(deep=True)
solo_out = austin_outcomes[~austin_outcomes['Animal ID'].isin(repeat_out['Animal ID'].unique())].copy(deep=True)

solo_in.shape[0] - solo_in['Animal ID'].isin(solo_out['Animal ID']).sum()

Missing outcome data on only 440 dogs that appeared once in the sets

### Merge Repeats

Probably only missing a few observations. Likely dogs in the system without outcomes at the time the data was collected

repeat_in.shape[0] - repeat_out.shape[0]

Drop redundant cols

repeat_out.drop(columns=['Name', 'Breed', 'Color'], inplace=True)

Rename columns

repeat_out.rename(columns={
    'Animal ID': 'Animal ID',
    'DateTime': 'date_out',
    'Date of Birth': 'dob',
    'Outcome Type': 'outcome',
    'Outcome Subtype': 'outcome_sub',
    'Sex upon Outcome': 'sex_out',
    'Age upon Outcome': 'age_out'
}, inplace=True)

Sort & merge repeat visitors

# storage
in_hold = repeat_in.copy()
in_hold.set_index('Animal ID', inplace=True)
out_hold = repeat_out.copy()                                                                
out_hold.set_index('Animal ID', inplace=True)

# drop missing IDs
in_hold.drop(list(set(in_hold.index).difference(set(out_hold.index))), inplace=True)
out_hold.drop(list(set(out_hold.index).difference(set(in_hold.index))), inplace=True)

# iterate
for i in repeat_in['Animal ID'].unique():                                                   # select ids of intake animals
    if i in set(repeat_out['Animal ID']):
        ins = repeat_in[repeat_in['Animal ID'] == i]                                        # grab intake rows
        outs = repeat_out[repeat_out['Animal ID'] == i]                                     # grab outcome rows

        if ins.shape[0] != outs.shape[0]:                                                   # check if ins & outs match up
            if i in set(in_hold.index):
                in_hold.drop(i, inplace=True)
            if i in set(out_hold.index):
                out_hold.drop(i, inplace=True)

repeat = pd.merge(left=in_hold, right=out_hold, how='left', on='Animal ID')
repeat.reset_index(drop=False, inplace=True)

Inspect repeat visitors

repeat[(repeat['date_out'] - repeat['DateTime']) < pd.Timedelta(0, units='minute')].head(7)

Looks like we have ~17,000 observations of repeat animal IDs that show up in system and do not have a clear intake date vs outcome date. Will drop these observations so they do not muddy our predictions.

(repeat[(repeat['date_out'] - repeat['DateTime']) > pd.Timedelta(0, units='minute')]['date_out'] - repeat[(repeat['date_out'] - repeat['DateTime']) > pd.Timedelta(0, units='minute')]['DateTime']).mean()

Average time for a repeat animal ID is 171 days. We will see later that the average time a dog is in the system that has only been in the system once is only 17 days. We're clearly having merging issues. If the original data had unique visit IDs, we could resolve this easily, but because time is a valuable resource, we will sadly have to drop the observations.

### Merge Solos

Drop missing outcomes for one time visitors

solo_in = solo_in[solo_in['Animal ID'].isin(solo_out['Animal ID'])].copy(deep=True)

Drop redundant columns

solo_out.drop(columns=['Name', 'Breed', 'Color'], inplace=True)

Rename columns

solo_out.rename(columns={
    'Animal ID': 'Animal ID',
    'DateTime': 'date_out',
    'Date of Birth': 'dob',
    'Outcome Type': 'outcome',
    'Outcome Subtype': 'outcome_sub',
    'Sex upon Outcome': 'sex_out',
    'Age upon Outcome': 'age_out'
}, inplace=True)

Merge intakes and outcomes

solo = pd.merge(left=solo_in, right=solo_out, how='left', on='Animal ID')

Inspect solo merge

solo[(solo['DateTime'] > solo['date_out'])].head()

None of the animals with a date out that is identical to their date in were in the system for longer than 1 day.

((solo[(solo['DateTime'] > solo['date_out'])])['date_out'] - (solo[(solo['DateTime'] > solo['date_out'])]['DateTime']) > pd.Timedelta(-1, units='days')).sum()

Will hold these in the frame for now, but will consider dropping later.

(solo[(solo['DateTime'] > solo['date_out'])])['date_out'] - (solo[(solo['DateTime'] > solo['date_out'])])['DateTime']

Check for correct time syncing

solo[solo['DateTime'] < solo['date_out']]

~55,000 observations

(solo[(solo['date_out'] - solo['DateTime']) > pd.Timedelta(0, units='minute')]['date_out'] - solo[(solo['date_out']   -   solo['DateTime']) > pd.Timedelta(0, units='minute')]['DateTime']).mean()

Average time spent in the shelter for dogs with only one appearance is ~16 days. Repeat dogs had an average significantly higher. Leads me to believe there are some serious inconsistencies in the data. Cannot reliably use these data points.

### Drop Repeats

Get unique IDs

solo_in = set(austin_intakes['Animal ID'].value_counts()[austin_intakes['Animal ID'].value_counts() == 1].index)
solo_out = set(austin_outcomes['Animal ID'].value_counts()[austin_outcomes['Animal ID'].value_counts() == 1].index)

Confirm IDs have both an intake and an outcome observation

solo_ids = solo_in.intersection(solo_out)

Filter missing

austin_intakes = austin_intakes[austin_intakes['Animal ID'].isin(solo_ids)]
austin_intakes.set_index('Animal ID', inplace=True)
austin_outcomes = austin_outcomes[austin_outcomes['Animal ID'].isin(solo_ids)]
austin_outcomes.set_index('Animal ID', inplace=True)

austin_intakes.shape

austin_outcomes.shape

Rename columns

austin_intakes.rename(columns={
    'Name': 'name',
    'DateTime': 'date_in',
    'Found Location': 'location',
    'Intake Type': 'intake_type',
    'Intake Condition': 'condition',
    'Sex upon Intake': 'sex_in',
    'Age upon Intake': 'age_in',
    'Breed': 'breed',
    'Color': 'color'
}, inplace=True)

austin_outcomes.rename(columns={
    'Name': 'name',
    'DateTime': 'date_out',
    'Date of Birth': 'dob',
    'Outcome Type': 'outcome',
    'Outcome Subtype': 'outcome_sub',
    'Sex upon Outcome': 'sex_out',
    'Age upon Outcome': 'age_out',
    'Breed': 'breed',
    'Color': 'color'
}, inplace=True)

austin_outcomes.drop(columns=['name', 'breed', 'color'], inplace=True)

## Final Merge

austin = pd.merge(left=austin_intakes, right=austin_outcomes, how='left', left_on='Animal ID', right_on='Animal ID')

austin.reset_index(drop=False, inplace=True)

austin.rename(columns={'Animal ID':'id'}, inplace=True)

Drop NA

austin = austin[austin['outcome'].notna()].copy(deep=True)

### Impute Missing

austin['sex_in'].value_counts()

austin['sex_in'].fillna('Intact Male', inplace=True)

Filling with most frequent

austin['sex_out'].value_counts()

austin['sex_out'].fillna('Neutered Male', inplace=True)

Filling with most frequent

austin['age_out'].value_counts()[:5]

austin['age_out'].fillna('1 year', inplace=True)

austin.reset_index(drop=True, inplace=True)

Convert month to decimil

austin.loc[austin[austin['age_out'].str.contains('month')].index.values, 'age_out'] = austin['age_out'][austin['age_out'].str.contains('month')].apply(lambda x: str(int(x.split()[0])/12))
austin.loc[austin[austin['age_in'].str.contains('month')].index.values, 'age_in'] = austin['age_in'][austin['age_in'].str.contains('month')].apply(lambda x: str(int(x.split()[0])/12))

Convert week to decimil

austin.loc[austin[austin['age_out'].str.contains('week')].index.values, 'age_out'] = austin['age_out'][austin['age_out'].str.contains('week')].apply(lambda x: str(int(x.split()[0])/52))
austin.loc[austin[austin['age_in'].str.contains('week')].index.values, 'age_in'] = austin['age_in'][austin['age_in'].str.contains('week')].apply(lambda x: str(int(x.split()[0])/52))

Convert day to decimil

austin.loc[austin[austin['age_out'].str.contains('day')].index.values, 'age_out'] = austin['age_out'][austin['age_out'].str.contains('day')].apply(lambda x: str(int(x.split()[0])/365))
austin.loc[austin[austin['age_in'].str.contains('day')].index.values, 'age_in'] = austin['age_in'][austin['age_in'].str.contains('day')].apply(lambda x: str(int(x.split()[0])/365))

Drop 'years' string

austin.loc[austin[austin['age_out'].str.contains('year')].index.values, 'age_out'] = austin['age_out'][austin['age_out'].str.contains('year')].apply(lambda x: str(x.split()[0]))
austin.loc[austin[austin['age_in'].str.contains('year')].index.values, 'age_in'] = austin['age_in'][austin['age_in'].str.contains('year')].apply(lambda x: str(x.split()[0]))

Fix negative ages

austin['age_out'] = austin['age_out'].str.replace('-', '')
austin['age_in'] = austin['age_in'].str.replace('-', '')

austin['age_out'] = round(austin['age_out'].astype(float), 2)
austin['age_in'] = round(austin['age_in'].astype(float), 2)

Binarize neuter/spay status

austin['sex_in'].value_counts()

austin['sex'] = austin['sex_in'].map({
    'Intact Male': 1,
    'Neutered Male': 1,
    'Intact Female': 0,
    'Spayed Female': 0,
    'Unknown': 1
})

austin['intact_in'] = austin['sex_in'].map({
    'Intact Male': 1,
    'Neutered Male': 0,
    'Intact Female': 1,
    'Spayed Female': 0,
    'Unknown': 1
})

austin['intact_out'] = austin['sex_out'].map({
    'Intact Male': 1,
    'Neutered Male': 0,
    'Intact Female': 1,
    'Spayed Female': 0,
    'Unknown': 1
})

Drop redundant cols

austin.drop(columns=['sex_in', 'sex_out'], inplace=True)

Convert Date of Birth to datetime

austin['dob'] = pd.to_datetime(austin['dob'])

Reinspect

austin.head(3)

austin.info()

## Intake, Outcome, and Condition Classes

### Intake Subtype

austin['intake_type'].value_counts()

Rename

austin['intake_type'] = austin['intake_type'].map(
    {
        'Owner Surrender':'surrender', 
        'Public Assist':'public_assist', 
        'Stray':'stray', 
        'Euthanasia Request':'euth_request',
        'Abandoned':'abandoned',
        'wildlife':'wildlife',
        }
    )

### Outcome Subtype

austin['outcome_sub'].isna().sum()

Too many missing values. Will drop.

austin.drop(columns=['outcome_sub'], inplace=True)

### Condition

austin['condition'].value_counts()

austin['condition'] = austin['condition'].map(
    {
        'Normal':'normal', 
        'Injured':'injured', 
        'Sick':'sick', 
        'Nursing':'nursing',
        'Aged': 'aged',
        'Other': 'other',
        'Medical': 'medical',
        'Pregnant': 'pregnant',
        'Neonatal': 'neonatal',
        'Behavior': 'behavior',
        'Med Urgent':'injured',
        'Med Attn ':'injured',
        }
    )

### Outcome

austin['outcome'].value_counts()

Rename

austin['outcome'] = austin['outcome'].map(
    {
        'Adoption': 'adoption', 
        'Transfer': 'transfer', 
        'Return to Owner': 'return_owner', 
        'Euthanasia': 'euthanasia',
        'Rto-Adopt': 'return_owner',
        'Died': 'died',
        'Disposal': 'disposal',
        'Missing': 'missing',
        }
    )

## Age Classifications

Will need to sync the Austin Animal Shelter age as a value into the same format as is passed with PetFinder API. We will lose a little context with our data, but at least we will have apples to apples to work with in the model.

austin['age_in'] = np.round(austin['age_in'])

austin['age_in'].value_counts()

austin['age_out'] = np.round(austin['age_out'])
austin['age_out'].value_counts()

(A few dogs have lived nice full lives. Happy for their doggo parents!!)

Add a column for our age classification.

age_cats = {
    0: 'Baby',
    1: 'Young',
    2: 'Adult', 3: 'Adult', 4: 'Adult', 5: 'Adult', 6: 'Adult', 7: 'Adult', 8: 'Adult',
    9: 'Senior', 10: 'Senior', 11: 'Senior', 12: 'Senior', 13: 'Senior', 14: 'Senior', 15: 'Senior', 16: 'Senior', 
    17: 'Senior', 18: 'Senior', 19: 'Senior', 20: 'Senior', 21: 'Senior', 22: 'Senior', 23: 'Senior', 24: 'Senior', 
    25: 'Senior', 26: 'Senior', 27: 'Senior', 28: 'Senior', 29: 'Senior', 30: 'Senior',
    np.nan: 'Adult' 
}

austin['age'] = austin['age_out'].map(age_cats)

austin['age'].value_counts()

## Colors Classification

368 unique colors is far too many. Will attempt to reduce the number of unique values.

austin['color'].value_counts()

Unique list of dog colors

colors = list(set(austin['color']))

Build primary color & secondary color combos and separate them into columns

primary_colors = []
secondary_colors = []

for i in range(len(austin['color'])):
    if str(austin['color'][i])!='nan':
        color_list = re.split(' |/', austin['color'][i])
        if len(color_list)==1:
            primary_colors.append(color_list[0])
            secondary_colors.append(color_list[0])
        else:
            primary_colors.append(color_list[0])
            secondary_colors.append(color_list[1])
    else:
        primary_colors.append(np.nan)
        secondary_colors.append(np.nan)   

austin['primary_color'] = primary_colors

austin['secondary_color'] = secondary_colors

austin.head(3)

Lowercase color columns

austin['primary_color'] = austin['primary_color'].str.lower()

Check for missing..

austin['primary_color'].isna().sum()

austin['primary_color'].value_counts()[:10]

austin['secondary_color'] = austin['secondary_color'].str.lower()

austin['secondary_color'].isna().sum()

austin['secondary_color'].value_counts()[:10]

# Export data

austin.to_csv('./datasets/working_data/austin.csv', index=False)

[Run Next](https://github.com/gwoodstock/gooddogs/blob/main/2_data_clean_shelter_breeds.ipynb): Breed data cleaning
