==> 2_data_clean_shelter_breeds.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np
import re

# Import Data

df = pd.read_csv('./datasets/working_data/austin.csv')

# Inspect Data

df.head(3)

df['breed'].nunique()

# Clean Breeds

2219 unique breeds!! That is far too many. Will need to reduce for modeling.

# lowercase and strip extra white space for breed names
df['breed'] = df['breed'].str.lower().str.strip()

# remove 'mix' at the end of the breed name
df['breed'] = df['breed'].str.replace(' mix', '')

# expand pitbull to pit bull for consistency
df['breed'].replace('pitbull', 'pit bull', inplace=True)

Identify unique singular breed observations

df[df['breed'].str.contains('/') == False]['breed'].nunique()

202 unique combo breeds. Will want to separate these into their own column and reduce wherever possible.

df[df['breed'].str.contains('/')].head(3)

Identify unique combo breed observations

df[df['breed'].str.contains('/')]['breed'].nunique()

Was able to reduce the total number of unique breeds by about 400. Still a lot of work to be done.

df['breed'].value_counts(ascending=True)[df['breed'].value_counts(ascending=True) == 1]

## Initial breed name cleaning

Regex to replace breed names to remain consistent

df['breed'] = df['breed'].map(lambda x: re.sub('labrador retr\Z', 'labrador retriever', x))
df['breed'] = df['breed'].map(lambda x: re.sub('germ\sshepherd', 'german shepherd', x))
df['breed'] = df['breed'].map(lambda x: re.sub('^staffordshire$', 'american staffordshire terrier', x))
df['breed'] = df['breed'].map(lambda x: re.sub('chihuahua shorthair', 'chihuahua', x))
df['breed'] = df['breed'].map(lambda x: re.sub('chihuahua longhair', 'chihuahua', x))
df['breed'] = df['breed'].map(lambda x: re.sub('chihuahua sh', 'chihuahua', x))
df['breed'] = df['breed'].map(lambda x: re.sub('chihuahua lh', 'chihuahua', x))
df['breed'] = df['breed'].map(lambda x: re.sub('airedale terr\Z', 'airedale terrier', x))
df['breed'] = df['breed'].map(lambda x: re.sub('amer ', 'american', x))
df['breed'] = df['breed'].map(lambda x: re.sub('alask ', 'alaskan', x))
df['breed'] = df['breed'].map(lambda x: re.sub('am pit bull ter', 'pit bull', x))
df['breed'] = df['breed'].map(lambda x: re.sub('aust ', 'australian', x))
df['breed'] = df['breed'].map(lambda x: re.sub('american bulldog', 'bulldog', x))
df['breed'] = df['breed'].map(lambda x: re.sub('english bulldog', 'bulldog', x))
df['breed'] = df['breed'].map(lambda x: re.sub('black/tan hound', 'black and tan coonhound', x))
df['breed'] = df['breed'].map(lambda x: re.sub('yorkshire terr\Z', 'yorkshire terrier', x))
df['breed'] = df['breed'].map(lambda x: re.sub('australiancattle dog', 'australian cattle dog', x))
df['breed'] = df['breed'].map(lambda x: re.sub('anatol shepherd', 'anatolian sheepdog', x))
df['breed'] = df['breed'].map(lambda x: re.sub('doberman pinsch\Z', 'doberman pinscher', x))
df['breed'] = df['breed'].map(lambda x: re.sub('poodle min\Z', 'miniature poodle', x))
df['breed'] = df['breed'].map(lambda x: re.sub('alaskan husky', 'siberian husky', x))
df['breed'] = df['breed'].map(lambda x: re.sub('rat terrier', 'fox terrier', x))
df['breed'] = df['breed'].map(lambda x: re.sub('wire hair fox terrier', 'fox terrier', x))
df['breed'] = df['breed'].map(lambda x: re.sub('americanbulldog', 'bulldog', x))
df['breed'] = df['breed'].map(lambda x: re.sub('australianshepherd', 'australian shepherd', x))
df['breed'] = df['breed'].map(lambda x: re.sub('dachshund wirehair', 'dachshund', x))
df['breed'] = df['breed'].map(lambda x: re.sub('american staff\Z', 'pit bull', x))
df['breed'] = df['breed'].map(lambda x: re.sub('american staffordshire terrier', 'pit bull', x))
df['breed'] = df['breed'].map(lambda x: re.sub('dachshund longhair', 'dachshund', x))
df['breed'] = df['breed'].map(lambda x: re.sub('schnauzer min\Z', 'miniature schnauzer', x))
df['breed'] = df['breed'].map(lambda x: re.sub('flat coat retriever', 'flat-coated retriever', x))
df['breed'] = df['breed'].map(lambda x: re.sub('collie smooth', 'collie', x))
df['breed'] = df['breed'].map(lambda x: re.sub('chinese sharpei', 'chinese shar pei', x))
df['breed'] = df['breed'].map(lambda x: re.sub('queensland heeler', 'australian cattle dog', x))
df['breed'] = df['breed'].map(lambda x: re.sub('rhod ridgeback', 'rhodesian ridgeback', x))
df['breed'] = df['breed'].map(lambda x: re.sub('german shorthair pointer', 'german shorthaired pointer', x))
df['breed'] = df['breed'].map(lambda x: re.sub('manchester terrier', 'standard manchester terrier', x))
df['breed'] = df['breed'].map(lambda x: re.sub('collie rough', 'collie', x))
df['breed'] = df['breed'].map(lambda x: re.sub('golden retr\Z', 'golden retriever', x))
df['breed'] = df['breed'].map(lambda x: re.sub('soft coated wheaten terrier', 'soft-coated wheaten terrier', x))
df['breed'] = df['breed'].map(lambda x: re.sub('west highland\Z', 'west highland white terrier', x))
df['breed'] = df['breed'].map(lambda x: re.sub('australian kelpie', 'australian cattle dog', x))
df['breed'] = df['breed'].map(lambda x: re.sub('redbone hound', 'redbone coonhound', x))
df['breed'] = df['breed'].map(lambda x: re.sub('eng bulldog', 'bulldog', x))

Import breed info for all dogs, convert it to a dictionary for more filtering.

all_dogs = pd.read_csv('./datasets/working_data/all_breeds_info.csv')
pure_dogs = pd.read_csv('./datasets/raw_data/dog_intelligence.csv')
master_list = set(list(all_dogs['breed']) + list(pure_dogs['Breed'].str.lower()))
all_dogs = all_dogs.to_dict(orient='records')

df.head(3)

New breed columns

df['breed_1'] = np.nan
df['breed_2'] = np.nan
df['pure'] = np.nan

Collect all the known breeds and compile in a new column.

def breed_check(x):
    if x in master_list:
        return x

df['breed_1'] = df['breed'].apply(breed_check)
df['breed_2'] = df['breed'].apply(breed_check)
df['pure'] = np.where(df['breed_1'] == df['breed_2'], 1, np.nan)

Observations with breed not missing from breed library.

df['breed_1'].isna().sum()

Most common observation breed names not in main breed library.

df[df['breed_1'].isna()]['breed'].value_counts()[:10]

Parse breed names if they contain a '/' and attempt to catch all known breed names in our new split breed columns.

# loops over all breeds in dataframe
df_mini = df[(df['breed_1'].isna()) & (df['breed'].str.contains('/'))].copy()
for i, dog in df_mini['breed'].items():

    # splits the dogs on the '/'
    breeds = dog.split('/')

    # checks if any items of breeds is in the master list of dog breeds
    if not set(breeds).isdisjoint(master_list):
        
        # checks both position combinations
        if (breeds[0] in master_list) and (breeds[1] in master_list):
            df.loc[i, 'breed_1'] = breeds[0]
            df.loc[i, 'breed_2'] = breeds[1]

Missing breeds have been reduced from just under 11,000 down to only 2,500!! Almost there.

df['breed_1'].isna().sum()

Make classifier for 'mixed' breeds

for i, v in df[df['breed'].str.contains('/')]['breed'].items():
    v = v.split('/')
    if v[0] in master_list and not v[1] in master_list:
        df.loc[i, 'breed_1'] = v[0]
        df.loc[i, 'breed_2'] = 'mix'
    elif not v[0] in master_list and v[1] in master_list:
        df.loc[i, 'breed_1'] = v[1]
        df.loc[i, 'breed_2'] = 'mix'

df['breed_1'].isna().sum()

Reduced missing breeds by 700

List remaining missing breeds. Probably best to drop these values.

set(df[df['breed_1'].isna()]['breed']).difference(master_list)

List most common

df[df['breed_1'].isna()]['breed'].value_counts()[:20]

# Export Data

df.to_csv('./datasets/working_data/austin_clean_breed.csv', index=False)

[Run Next](https://github.com/gwoodstock/gooddogs/blob/main/3_data_clean_petfinder.ipynb): PetFinder API data cleaning
