==> 8_modela_rfc_stripped.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np
import pickle

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import OneHotEncoder

# Feature Selection

target = ['at_risk']
features = ['age', 'breed_1', 'condition', 'primary_color']
# filters = ['intake_type', 'condition', 'outcome']

# Import Data

data = pd.read_csv('./datasets/working_data/model_data_full.csv', usecols=features + target) # + filters)

data.head(2)

# Filters

data.rename(columns={
    'condition': 'cond',
    'breed_1': 'breed',
    'at_risk': 'target',
    'primary_color': 'color'
}, inplace=True)

data.head(2)

data.isna().sum()

data['cond'] = np.where(data['cond'] == 'normal', 1, 0)

# Export Data

data.to_csv('./datasets/cleaned_data/model_data.csv', index=False)

# Model

y = data['target']
X = data.drop('target', axis=1)

ohe = OneHotEncoder(handle_unknown='ignore').fit(np.array(X[['age', 'breed', 'cond']]))
X = ohe.transform(np.array(X[['age', 'breed', 'cond']]))

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=123, stratify=y)

rfc = RandomForestClassifier(n_estimators=300, max_depth=200, random_state=123, min_samples_split=10)
rfc.fit(X_train, y_train)

# Score

print(rfc.score(X_train, y_train))
print(rfc.score(X_test, y_test))

Save the model and encoder for the web app.

with open("./models/model_rfc.pkl", "wb") as file:
    pickle.dump(rfc, file)
with open("./models/encode_data.pkl", "wb") as file:
    pickle.dump(ohe, file)

[Run Next](https://github.com/gwoodstock/gooddogs/blob/main/9_models_GPU.ipynb): XGB & Neural Net.
