==> 11_predict_petfinder.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np
import pickle

from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import OneHotEncoder

# Import Data

data = pd.read_csv('./datasets/cleaned_data/petfinder.csv')

data.head()

# Load Models

# load model
with open('./models/model_rfc.pkl', 'rb') as f:
    model = pickle.load(f)

# load transformer
with open('./models/encode_data.pkl', 'rb') as f:
    encoder = pickle.load(f)

# Data Cleaning

data['cond'] = data['injured'].map(lambda x: np.where(x == 1, 0, 1))
data['age'] = data['age'].str.lower()
data['breed'] = data['breed_1'].str.lower()

X = np.array(data[['age', 'breed', 'cond']])
X = encoder.transform(X)

preds = model.predict_proba(X)

preds[:10, 1]

data['pred_at_risk'] = preds[:,1]

# Export Predictions

data.to_csv('./datasets/cleaned_data/petfinder.csv', index=False)

We will use these predictions on the web portal in a [Tableau Dashboard](https://public.tableau.com/app/profile/anna.rutledge4407/viz/FindyourDog/Dashboard1) that will allow users to select various conditions and see how the mean predictions react. Check them out for yourself!
